LPS-206586 Remove PortalCacheClusterChannel internal interface abstraction
